# apps/server/swagger/swagger.yaml
basePath: /api
definitions:
  internal_modules_analytics.CreatePollRequest:
    description: Request body for creating a poll on a problem
    properties:
      problemId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      question:
        example: How difficult did you find this problem?
        maxLength: 240
        minLength: 10
        type: string
    required:
    - problemId
    - question
    type: object
...
  internal_modules_analytics.LeaderboardEntryResponse:
    description: Response containing a single leaderboard entry
    properties:
      data:
        $ref: '#/definitions/internal_modules_analytics.LeaderboardEntryData'
      success:
        example: true
        type: boolean
    type: object
  internal_modules_analytics.LeaderboardResponse:
    description: Response containing leaderboard entries with pagination
    properties:
      data:
        items:
          $ref: '#/definitions/internal_modules_analytics.LeaderboardEntryData'
        type: array
      meta:
        $ref: '#/definitions/internal_modules_analytics.OffsetPagination'
      success:
        example: true
        type: boolean
    type: object
  internal_modules_analytics.OffsetPagination:
    description: Offset-based pagination metadata
    properties:
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  internal_modules_analytics.PollListResponse:
    description: Response containing a list of polls with pagination
    properties:
      data:
        items:
          $ref: '#/definitions/internal_modules_analytics.PollData'
        type: array
      meta:
        $ref: '#/definitions/internal_modules_analytics.OffsetPagination'
      success:
        example: true
        type: boolean
    type: object
  internal_modules_analytics.PollResponse:
    description: Response containing a single poll
    properties:
      data:
        $ref: '#/definitions/internal_modules_analytics.PollData'
      success:
        example: true
        type: boolean
    type: object
...
  internal_modules_analytics.PollResultsResponse:
    description: Response containing aggregated poll results
    properties:
      data:
        $ref: '#/definitions/internal_modules_analytics.PollResultsData'
      success:
        example: true
        type: boolean
    type: object
...
  internal_modules_analytics.VotePollRequest:
    description: Request body for casting or updating a vote on a poll
    properties:
      vote:
        enum:
        - easy
        - medium
        - hard
        example: medium
        type: string
    required:
    - vote
    type: object
  internal_modules_analytics.VoteResponse:
    description: Response containing a single vote
    properties:
      data:
        $ref: '#/definitions/internal_modules_analytics.VoteData'
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@coderz.space
    name: API Support
  description: Comprehensive bootcamp management platform API with multi-tenant architecture
    and role-based access control
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Coderz.space Bootcamp Management API
  version: "1.0"
paths:
...
  /v1/bootcamps/{bootcampId}/leaderboard:
    get:
      consumes:
      - application/json
      description: Retrieve pre-calculated leaderboard rankings for a bootcamp. Returns
        snapshot data without real-time recalculation. User must be enrolled in the
        bootcamp.
      parameters:
      - description: Bootcamp ID (UUID)
        in: path
        name: bootcampId
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Leaderboard entries with pagination
          schema:
            $ref: '#/definitions/internal_modules_analytics.LeaderboardResponse'
        "400":
          description: Bad request - invalid bootcamp ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - not enrolled in bootcamp
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found - bootcamp does not exist
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get bootcamp leaderboard
      tags:
      - Leaderboards
  /v1/bootcamps/{bootcampId}/leaderboard/{enrollmentId}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific leaderboard entry by enrollment ID. Mentees
        can only view their own entry.
      parameters:
      - description: Bootcamp ID (UUID)
        in: path
        name: bootcampId
        required: true
        type: string
      - description: Bootcamp Enrollment ID (UUID)
        in: path
        name: enrollmentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Leaderboard entry details
          schema:
            $ref: '#/definitions/internal_modules_analytics.LeaderboardEntryResponse'
        "400":
          description: Bad request - invalid ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - access denied
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found - entry does not exist
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get leaderboard entry
      tags:
      - Leaderboards
  /v1/bootcamps/{bootcampId}/polls:
    get:
      consumes:
      - application/json
      description: List polls for a bootcamp with optional problem filtering. Includes
        user's vote if they have voted. User must be enrolled in bootcamp.
      parameters:
      - description: Bootcamp ID (UUID)
        in: path
        name: bootcampId
        required: true
        type: string
      - description: Filter by problem ID (UUID)
        in: query
        name: problemId
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of polls with pagination
          schema:
            $ref: '#/definitions/internal_modules_analytics.PollListResponse'
        "400":
          description: Bad request - invalid bootcamp ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - not enrolled in bootcamp
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List polls
      tags:
      - Polls
    post:
      consumes:
      - application/json
      description: Create a difficulty poll for a problem in a bootcamp (mentor/admin
        only). Supports idempotency via Idempotency-Key header.
      parameters:
      - description: Bootcamp ID (UUID)
        in: path
        name: bootcampId
        required: true
        type: string
      - description: Poll details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_analytics.CreatePollRequest'
      - description: Idempotency key for safe retries
        in: header
        name: Idempotency-Key
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Poll created successfully
          schema:
            $ref: '#/definitions/internal_modules_analytics.PollResponse'
        "400":
          description: Bad request - validation error or invalid problem ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - mentor/admin role required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found - problem does not exist
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a poll
      tags:
      - Polls
  /v1/bootcamps/{bootcampId}/polls/{pollId}:
    get:
      consumes:
      - application/json
      description: Retrieve full details of a specific poll including user's vote
        state. User must be enrolled in bootcamp.
      parameters:
      - description: Bootcamp ID (UUID)
        in: path
        name: bootcampId
        required: true
        type: string
      - description: Poll ID (UUID)
        in: path
        name: pollId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Poll details
          schema:
            $ref: '#/definitions/internal_modules_analytics.PollResponse'
        "400":
          description: Bad request - invalid poll ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - not enrolled in bootcamp
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found - poll does not exist
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get poll details
      tags:
      - Polls
  /v1/bootcamps/{bootcampId}/polls/{pollId}/results:
    get:
      consumes:
      - application/json
      description: Retrieve aggregated poll results with vote counts and percentages
        (mentor/admin/super_admin only). Mentees cannot access results.
      parameters:
      - description: Bootcamp ID (UUID)
        in: path
        name: bootcampId
        required: true
        type: string
      - description: Poll ID (UUID)
        in: path
        name: pollId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Aggregated poll results
          schema:
            $ref: '#/definitions/internal_modules_analytics.PollResultsResponse'
        "400":
          description: Bad request - invalid poll ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - mentor/admin/super_admin role required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found - poll does not exist
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get poll results
      tags:
      - Polls
  /v1/bootcamps/{bootcampId}/polls/{pollId}/vote:
    put:
      consumes:
      - application/json
      description: Cast or update a vote on a poll (mentee only). Uses PUT method
        for idempotent vote creation/update. Returns 201 for first vote, 200 for updates.
      parameters:
      - description: Bootcamp ID (UUID)
        in: path
        name: bootcampId
        required: true
        type: string
      - description: Poll ID (UUID)
        in: path
        name: pollId
        required: true
        type: string
      - description: Vote details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_analytics.VotePollRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Vote updated successfully
          schema:
            $ref: '#/definitions/internal_modules_analytics.VoteResponse'
        "201":
          description: Vote created successfully
          schema:
            $ref: '#/definitions/internal_modules_analytics.VoteResponse'
        "400":
          description: Bad request - validation error or invalid poll ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - only mentees can vote
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found - poll does not exist
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Vote on a poll
      tags:
      - Polls
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...